feat: Start customizing the template
https://harperdb.atlassian.net/browse/STUDIO-551
